% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{importance.divfor}
\alias{importance.divfor}
\alias{importance}
\title{Diversity Forest variable importance}
\usage{
\method{importance}{divfor}(x, ...)
}
\arguments{
\item{x}{\code{divfor} object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Variable importance measures.
}
\description{
Extract variable importance of \code{divfor} object.
}
\seealso{
\code{\link{divfor}}
}
\author{
Marvin N. Wright
}
