% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{predictions.divfor}
\alias{predictions.divfor}
\title{Diversity Forest predictions}
\usage{
\method{predictions}{divfor}(x, ...)
}
\arguments{
\item{x}{\code{divfor} object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Predictions: Classes for Classification forests, Numerical values for Regressions forests and the estimated survival functions for all individuals for Survival forests.
}
\description{
Extract training data predictions of \code{divfor} object.
}
\seealso{
\code{\link{divfor}}
}
\author{
Marvin N. Wright
}
