% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.divfor.forest}
\alias{print.divfor.forest}
\title{Print divfor forest}
\usage{
\method{print}{divfor.forest}(x, ...)
}
\arguments{
\item{x}{Object of class 'divfor.forest'.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print contents of divfor forest object.
}
\author{
Marvin N. Wright
}
