% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.divfor.prediction}
\alias{print.divfor.prediction}
\title{Print divfor prediction}
\usage{
\method{print}{divfor.prediction}(x, ...)
}
\arguments{
\item{x}{Object of class 'divfor.prediction'.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print contents of divfor prediction object.
}
\author{
Marvin N. Wright
}
