% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.divfor}
\alias{print.divfor}
\title{Print divfor}
\usage{
\method{print}{divfor}(x, ...)
}
\arguments{
\item{x}{Object of class 'divfor'.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print contents of divfor object.
}
\seealso{
\code{\link{divfor}}
}
\author{
Marvin N. Wright, Roman Hornung
}
