% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.tunedivfor}
\alias{print.tunedivfor}
\title{Print tunedivfor}
\usage{
\method{print}{tunedivfor}(x, ...)
}
\arguments{
\item{x}{Object of class 'tunedivfor'.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print contents of tunedivfor object.
}
\seealso{
\code{\link{tunedivfor}}
}
\author{
Roman Hornung
}
